refactor(coderd/mcp.go): inline oidcMCPTokenSource adapter
Drop coderd/mcp_oidc_token_source.go and put the small
UserOIDCTokenSource adapter in coderd/mcp.go alongside the rest of
the MCP server config code.
